Moved modules/contentItems/lib/testResources to app/lib/testResources, so that dummyContentItems can be used for testing elsewhere in the project as well.
