+++ app/modules/contentItems/components/EditableDisplay/typesToComponentsMap/Root.test.js
@@ -5,8 +5,9 @@ import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import { mount, shallow } from 'enzyme';
 
+import { dummyContentItemData as dummyData } from 'lib/testResources';
+
 import * as m from '../../../model';
-import { dummyContentItemData as dummyData } from '../../../lib/testResources';
 
 import { PureRoot } from './Root';
