+++ app/modules/contentItems/components/HtmlDisplay/index.test.js
@@ -3,8 +3,9 @@
 import * as React from 'react';
 import { mount, shallow } from 'enzyme';
 
+import { dummyContentItemData as dummyData } from 'lib/testResources';
+
 import * as m from '../../model';
-import { dummyContentItemData as dummyData } from '../../lib/testResources';
 
 import { PureHtmlDisplay } from '.';
