+++ app/modules/contentItems/components/HtmlDisplay/typesToComponentsMap/Heading.test.js
@@ -3,8 +3,9 @@
 import * as React from 'react';
 import { render, shallow } from 'enzyme';
 
+import { dummyContentItemData as dummyData } from 'lib/testResources';
+
 import * as m from '../../../model';
-import { dummyContentItemData as dummyData } from '../../../lib/testResources';
 
 import { PureHeading } from './Heading';
