+++ app/modules/contentItems/lib/edit/addChildOrSubItemIdToContext/index.test.js
@@ -2,9 +2,9 @@
 
 import { InvalidArgumentError, ObjectNotFoundError } from 'errors';
 import type { Identifier } from 'types/model';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as m from '../../../model';
-import { dummyContentItemData as dummyData } from '../../testResources';
 
 import edit from '..';
