+++ app/modules/contentItems/lib/find/allChildOrSubItems/index.test.js
@@ -3,9 +3,9 @@
 import _ from 'lodash';
 
 import { CorruptedInternalStateError } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as m from '../../../model';
-import { dummyContentItemData as dummyData } from '../../testResources';
 
 import find from '..';
