+++ app/modules/contentItems/reducer/addToState.test.js
@@ -1,10 +1,10 @@
 // @flow
 
 import { InvalidArgumentError, NotYetImplementedError, UnsupportedOperationError } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as t from '../actionTypes';
 import * as m from '../model';
-import { dummyContentItemData as dummyData } from '../lib/testResources';
 import edit from '../lib/edit';
 
 import reducer from '.';
