+++ app/modules/contentItems/reducer/moveInState.test.js
@@ -5,10 +5,10 @@ import {
   ObjectNotFoundError,
   UnsupportedOperationError,
 } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as t from '../actionTypes';
 import * as m from '../model';
-import { dummyContentItemData as dummyData } from '../lib/testResources';
 import edit from '../lib/edit';
 
 import reducer from '.';
