+++ app/modules/contentItems/reducer/setMultipleInState.test.js
@@ -1,8 +1,9 @@
 // @flow
 
+import { dummyContentItemData as dummyData } from 'lib/testResources';
+
 import * as t from '../actionTypes';
 import * as m from '../model';
-import { dummyContentItemData as dummyData } from '../lib/testResources';
 
 import reducer from '.';
