+++ app/modules/contentItems/reducer/switchEditingInState.test.js
@@ -1,10 +1,10 @@
 // @flow
 
 import { UnsupportedOperationError, ObjectNotFoundError } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as t from '../actionTypes';
 import * as m from '../model';
-import { dummyContentItemData as dummyData } from '../lib/testResources';
 
 import reducer from '.';
