+++ app/modules/contentItems/saga/task/edit.test.js
@@ -3,11 +3,11 @@
 import { expectSaga, testSaga } from 'redux-saga-test-plan';
 
 import { NotYetImplementedError, ObjectNotFoundError } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as t from '../../actionTypes';
 import * as m from '../../model';
 import selectors from '../../selectors';
-import { dummyContentItemData as dummyData } from '../../lib/testResources';
 
 import editSaga from './edit';
