+++ app/modules/contentItems/saga/task/reverseIndent.test.js
@@ -3,10 +3,10 @@
 import { expectSaga } from 'redux-saga-test-plan';
 
 import { ObjectNotFoundError } from 'errors';
+import { dummyContentItemData as dummyData } from 'lib/testResources';
 
 import * as t from '../../actionTypes';
 import * as m from '../../model';
-import { dummyContentItemData as dummyData } from '../../lib/testResources';
 
 import reverseIndentSaga from './reverseIndent';
