+++ app/modules/contentItems/selectors/getAll.test.js
@@ -1,7 +1,8 @@
 // @flow
 
+import { dummyContentItemData as dummyData } from 'lib/testResources';
+
 import * as m from '../model';
-import { dummyContentItemData as dummyData } from '../lib/testResources';
 
 import selectors from '.';
